backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
google-genai
sqlalchemy